TASK: Convert MenuToggler to typescript
